cos/tx_completion.rs: colocate park_reason + queue_state tests with their production code
Phase 1 of Path A test redistribution (third sibling PR to
refactor/tests-rings and refactor/tests-transmit). Move
`count_park_reason_helper_advances_exact_counter` and
`normalize_cos_queue_state_repairs_nonempty_unparked_queue_to_runnable`
from tx/mod.rs::tests into a fresh `#[cfg(test)] mod tests` at the
bottom of cos/tx_completion.rs.

The new tests block needs explicit imports for items that
tx_completion.rs does not glob in its production scope:
COS_MIN_BURST_BYTES (cos/token_bucket), TX_BATCH_SIZE
(crate::afxdp), and the `pub(super)` types
COS_FLOW_FAIR_BUCKETS, CoSQueueDropCounters, CoSQueueOwnerProfile,
FlowRrRing (crate::afxdp::types).

Mechanical: 2 tests, 87 lines moved. cargo test --bins 865/0/2.
